requirements.txt
fastapi
uvicorn
vertexai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv